Moved ColorConversion to separate assembly
