+++ .gitignore
@@ -110,3 +110,4 @@ UpgradeLog*.XML
 /src/Q42.HueApi/NuGet/Package/lib/Q42.HueApi.dll
 /src/Q42.HueApi/NuGet/*.nupkg
 *.nupkg
+/src/Q42.HueApi.ColorConverters/project.lock.json
