+++ src/Q42.HueApi.ColorConverters/OriginalWithModel/Extensions/LightExtensions.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Q42.HueApi.ColorConverters.OriginalWithModel
+{
+	public static class LightExtensions
+	{
+		public static string ToHex(this Light light)
+		{
+			return HueColorConverter.HexFromState(light.State, light.ModelId);
+		}
+	}
+}
