+++ src/Q42.HueApi.ColorConverters/project.json
@@ -1,0 +1,7 @@
+{
+  "supports": {},
+  "dependencies": {},
+  "frameworks": {
+    ".NETPortable,Version=v4.5,Profile=Profile259": {}
+  }
+}
