+++ src/Q42.HueApi.Tests/Q42.HueApi.Tests.csproj
@@ -67,8 +67,6 @@
     </Otherwise>
   </Choose>
   <ItemGroup>
-    <Compile Include="ColorConverterTests-Arrays.cs" />
-    <Compile Include="ColorConverterTests.cs" />
     <Compile Include="HueApiTests.cs" />
     <Compile Include="Remote\RemoteLightsTests.cs" />
     <Compile Include="RuleTests.cs" />
