+++ src/Q42.HueApi.WinRT.Sample/ViewModel/MainViewModel.cs
@@ -1,5 +1,7 @@
 using GalaSoft.MvvmLight;
 using GalaSoft.MvvmLight.Command;
+using Q42.HueApi;
+using Q42.HueApi.ColorConverters.OriginalWithModel;
 using Q42.HueApi.Interfaces;
 using Q42.WinRT.Data;
 using Q42.WinRT.Portable.Data;
