+++ src/Q42.HueApi.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
-VisualStudioVersion = 14.0.23107.0
+VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Q42.HueApi", "Q42.HueApi\Q42.HueApi.csproj", "{E59A16B4-6756-4576-AB63-6A9B2CF2892B}"
 EndProject
@@ -16,6 +16,10 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Q42.HueApi.NET.Tests", "Q42.HueApi.NET.Tests\Q42.HueApi.NET.Tests.csproj", "{2114B768-86CF-4687-BCFD-2B7916C85CD9}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Q42.HueApi.Tests", "Q42.HueApi.Tests\Q42.HueApi.Tests.csproj", "{0B96BC44-0006-46C2-BA96-CD63C123FD49}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Q42.HueApi.ColorConverters", "Q42.HueApi.ColorConverters\Q42.HueApi.ColorConverters.csproj", "{A4C0C8C0-DA94-4E12-A293-860B7CF609DB}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Q42.HueApi.ColorConverters.Tests", "Q42.HueApi.ColorConverters.Tests\Q42.HueApi.ColorConverters.Tests.csproj", "{64A5861F-4D4E-4127-B70F-F80CC9D49CFD}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -133,6 +137,38 @@ Global
 		{0B96BC44-0006-46C2-BA96-CD63C123FD49}.Release|ARM.ActiveCfg = Release|Any CPU
 		{0B96BC44-0006-46C2-BA96-CD63C123FD49}.Release|x64.ActiveCfg = Release|Any CPU
 		{0B96BC44-0006-46C2-BA96-CD63C123FD49}.Release|x86.ActiveCfg = Release|Any CPU
+		{A4C0C8C0-DA94-4E12-A293-860B7CF609DB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A4C0C8C0-DA94-4E12-A293-860B7CF609DB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A4C0C8C0-DA94-4E12-A293-860B7CF609DB}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{A4C0C8C0-DA94-4E12-A293-860B7CF609DB}.Debug|ARM.Build.0 = Debug|Any CPU
+		{A4C0C8C0-DA94-4E12-A293-860B7CF609DB}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{A4C0C8C0-DA94-4E12-A293-860B7CF609DB}.Debug|x64.Build.0 = Debug|Any CPU
+		{A4C0C8C0-DA94-4E12-A293-860B7CF609DB}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{A4C0C8C0-DA94-4E12-A293-860B7CF609DB}.Debug|x86.Build.0 = Debug|Any CPU
+		{A4C0C8C0-DA94-4E12-A293-860B7CF609DB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A4C0C8C0-DA94-4E12-A293-860B7CF609DB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A4C0C8C0-DA94-4E12-A293-860B7CF609DB}.Release|ARM.ActiveCfg = Release|Any CPU
+		{A4C0C8C0-DA94-4E12-A293-860B7CF609DB}.Release|ARM.Build.0 = Release|Any CPU
+		{A4C0C8C0-DA94-4E12-A293-860B7CF609DB}.Release|x64.ActiveCfg = Release|Any CPU
+		{A4C0C8C0-DA94-4E12-A293-860B7CF609DB}.Release|x64.Build.0 = Release|Any CPU
+		{A4C0C8C0-DA94-4E12-A293-860B7CF609DB}.Release|x86.ActiveCfg = Release|Any CPU
+		{A4C0C8C0-DA94-4E12-A293-860B7CF609DB}.Release|x86.Build.0 = Release|Any CPU
+		{64A5861F-4D4E-4127-B70F-F80CC9D49CFD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{64A5861F-4D4E-4127-B70F-F80CC9D49CFD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{64A5861F-4D4E-4127-B70F-F80CC9D49CFD}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{64A5861F-4D4E-4127-B70F-F80CC9D49CFD}.Debug|ARM.Build.0 = Debug|Any CPU
+		{64A5861F-4D4E-4127-B70F-F80CC9D49CFD}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{64A5861F-4D4E-4127-B70F-F80CC9D49CFD}.Debug|x64.Build.0 = Debug|Any CPU
+		{64A5861F-4D4E-4127-B70F-F80CC9D49CFD}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{64A5861F-4D4E-4127-B70F-F80CC9D49CFD}.Debug|x86.Build.0 = Debug|Any CPU
+		{64A5861F-4D4E-4127-B70F-F80CC9D49CFD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{64A5861F-4D4E-4127-B70F-F80CC9D49CFD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{64A5861F-4D4E-4127-B70F-F80CC9D49CFD}.Release|ARM.ActiveCfg = Release|Any CPU
+		{64A5861F-4D4E-4127-B70F-F80CC9D49CFD}.Release|ARM.Build.0 = Release|Any CPU
+		{64A5861F-4D4E-4127-B70F-F80CC9D49CFD}.Release|x64.ActiveCfg = Release|Any CPU
+		{64A5861F-4D4E-4127-B70F-F80CC9D49CFD}.Release|x64.Build.0 = Release|Any CPU
+		{64A5861F-4D4E-4127-B70F-F80CC9D49CFD}.Release|x86.ActiveCfg = Release|Any CPU
+		{64A5861F-4D4E-4127-B70F-F80CC9D49CFD}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
