+++ src/Q42.HueApi/Models/Light.cs
@@ -45,9 +45,6 @@ namespace Q42.HueApi
     [DataMember (Name = "swversion")]
     public string SoftwareVersion { get; set; }
 
-	public string ToHex()
-	{
-		return HueColorConverter.HexFromState(this.State, this.ModelId);
-	}
+	
   }
 }
