+++ src/Q42.HueApi/Models/LightCommand.cs
@@ -156,82 +156,14 @@ namespace Q42.HueApi
   /// </summary>
   public static class lightCommandExtensions
   {
-    /// <summary>
-    /// Helper to set the color based on a HEX value
-    /// </summary>
-    /// <param name="lightCommand"></param>
-    /// <param name="hexColor"></param>
-    /// <returns></returns>
-    public static LightCommand SetColor(this LightCommand lightCommand, string hexColor)
-    {
-      if (lightCommand == null)
-        throw new ArgumentNullException ("lightCommand");
-      if (hexColor == null)
-        throw new ArgumentNullException ("hexColor");
-     
-      //Clean hexColor value, remove the #
-      hexColor = hexColor.Replace("#", string.Empty);
-
-      if(hexColor.Length != 6)
-        throw new ArgumentException("hexColor should contains 6 characters", "hexColor");
-
-      int red = int.Parse(hexColor.Substring(0, 2), NumberStyles.AllowHexSpecifier);
-      int green = int.Parse(hexColor.Substring(2, 2), NumberStyles.AllowHexSpecifier);
-      int blue = int.Parse(hexColor.Substring(4, 2), NumberStyles.AllowHexSpecifier);
-
-      return lightCommand.SetColor(red, green, blue);
-    }
-
-    /// <summary>
-    /// Helper to set the color based on RGB strings
-    /// </summary>
-    /// <param name="lightCommand"></param>
-    /// <param name="red"></param>
-    /// <param name="green"></param>
-    /// <param name="blue"></param>
-    /// <returns></returns>
-    public static LightCommand SetColor(this LightCommand lightCommand, string red, string green, string blue)
-    {
-      if (lightCommand == null)
-        throw new ArgumentNullException ("lightCommand");
-
-      return lightCommand.SetColor(int.Parse(red), int.Parse(green), int.Parse(blue));
-    }
-
-    /// <summary>
-    /// Helper to set the color based on RGB
-    /// </summary>
-    /// <param name="lightCommand"></param>
-    /// <param name="red"></param>
-    /// <param name="green"></param>
-    /// <param name="blue"></param>
-    /// <returns></returns>
-    public static LightCommand SetColor(this LightCommand lightCommand, int red, int green, int blue, string model = "LCT001")
-    {
-      if (lightCommand == null)
-        throw new ArgumentNullException ("lightCommand");
-
-      var point = HueColorConverter.RgbToXY(new RGBColor(red / 255.0, green / 255.0, blue / 255.0), model);
-      return lightCommand.SetColor(point.x, point.y);
-    }
-
-	public static LightCommand SetColor(this LightCommand lightCommand, RGBColor color, string model)
-	{
-		if (lightCommand == null)
-			throw new ArgumentNullException("lightCommand");
-
-		var point = HueColorConverter.RgbToXY(color, model);
-		return lightCommand.SetColor(point.x, point.y);
-	}
-
-		/// <summary>
-		/// Helper to set the color based on the light's built in XY color schema
-		/// </summary>
-		/// <param name="lightCommand"></param>
-		/// <param name="x"></param>
-		/// <param name="y"></param>
-		/// <returns></returns>
-		public static LightCommand SetColor(this LightCommand lightCommand, double x, double y)
+	/// <summary>
+	/// Helper to set the color based on the light's built in XY color schema
+	/// </summary>
+	/// <param name="lightCommand"></param>
+	/// <param name="x"></param>
+	/// <param name="y"></param>
+	/// <returns></returns>
+	public static LightCommand SetColor(this LightCommand lightCommand, double x, double y)
     {
       if (lightCommand == null)
         throw new ArgumentNullException ("lightCommand");
